requirements.txt
tensorflow-gpu~=2.10.0
deepface~=0.0.69
matplotlib~=3.7.2
keras~=2.10.0
numpy~=1.24.3
opencv-python~=4.8.0.74